requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==3.2.2
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
diskcache==5.6.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
flake8==7.1.2
h11==0.14.0
httptools==0.6.4
idna==3.10
isort==6.0.0
Jinja2==3.1.5
llama_cpp_python==0.3.7
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlalchemy-stubs==0.4
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
